refactor: extract embedded data into a separate file for better organization
