+++ packages/webtrit_appearance_theme/lib/models/features_config/app_config.freezed.dart
@@ -3317,451 +3317,3 @@ abstract class _AppConfigSettingsItem extends AppConfigSettingsItem {
   _$$AppConfigSettingsItemImplCopyWith<_$AppConfigSettingsItemImpl>
       get copyWith => throw _privateConstructorUsedError;
 }
-
-EmbeddedData _$EmbeddedDataFromJson(Map<String, dynamic> json) {
-  return _EmbeddedData.fromJson(json);
-}
-
-/// @nodoc
-mixin _$EmbeddedData {
-  int? get id => throw _privateConstructorUsedError;
-  String get resource => throw _privateConstructorUsedError;
-  Map<String, dynamic> get attributes => throw _privateConstructorUsedError;
-  ToolbarConfig get toolbar => throw _privateConstructorUsedError;
-  Metadata get metadata => throw _privateConstructorUsedError;
-
-  /// Serializes this EmbeddedData to a JSON map.
-  Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
-
-  /// Create a copy of EmbeddedData
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  $EmbeddedDataCopyWith<EmbeddedData> get copyWith =>
-      throw _privateConstructorUsedError;
-}
-
-/// @nodoc
-abstract class $EmbeddedDataCopyWith<$Res> {
-  factory $EmbeddedDataCopyWith(
-          EmbeddedData value, $Res Function(EmbeddedData) then) =
-      _$EmbeddedDataCopyWithImpl<$Res, EmbeddedData>;
-  @useResult
-  $Res call(
-      {int? id,
-      String resource,
-      Map<String, dynamic> attributes,
-      ToolbarConfig toolbar,
-      Metadata metadata});
-
-  $ToolbarConfigCopyWith<$Res> get toolbar;
-  $MetadataCopyWith<$Res> get metadata;
-}
-
-/// @nodoc
-class _$EmbeddedDataCopyWithImpl<$Res, $Val extends EmbeddedData>
-    implements $EmbeddedDataCopyWith<$Res> {
-  _$EmbeddedDataCopyWithImpl(this._value, this._then);
-
-  // ignore: unused_field
-  final $Val _value;
-  // ignore: unused_field
-  final $Res Function($Val) _then;
-
-  /// Create a copy of EmbeddedData
-  /// with the given fields replaced by the non-null parameter values.
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? id = freezed,
-    Object? resource = null,
-    Object? attributes = null,
-    Object? toolbar = null,
-    Object? metadata = null,
-  }) {
-    return _then(_value.copyWith(
-      id: freezed == id
-          ? _value.id
-          : id // ignore: cast_nullable_to_non_nullable
-              as int?,
-      resource: null == resource
-          ? _value.resource
-          : resource // ignore: cast_nullable_to_non_nullable
-              as String,
-      attributes: null == attributes
-          ? _value.attributes
-          : attributes // ignore: cast_nullable_to_non_nullable
-              as Map<String, dynamic>,
-      toolbar: null == toolbar
-          ? _value.toolbar
-          : toolbar // ignore: cast_nullable_to_non_nullable
-              as ToolbarConfig,
-      metadata: null == metadata
-          ? _value.metadata
-          : metadata // ignore: cast_nullable_to_non_nullable
-              as Metadata,
-    ) as $Val);
-  }
-
-  /// Create a copy of EmbeddedData
-  /// with the given fields replaced by the non-null parameter values.
-  @override
-  @pragma('vm:prefer-inline')
-  $ToolbarConfigCopyWith<$Res> get toolbar {
-    return $ToolbarConfigCopyWith<$Res>(_value.toolbar, (value) {
-      return _then(_value.copyWith(toolbar: value) as $Val);
-    });
-  }
-
-  /// Create a copy of EmbeddedData
-  /// with the given fields replaced by the non-null parameter values.
-  @override
-  @pragma('vm:prefer-inline')
-  $MetadataCopyWith<$Res> get metadata {
-    return $MetadataCopyWith<$Res>(_value.metadata, (value) {
-      return _then(_value.copyWith(metadata: value) as $Val);
-    });
-  }
-}
-
-/// @nodoc
-abstract class _$$EmbeddedDataImplCopyWith<$Res>
-    implements $EmbeddedDataCopyWith<$Res> {
-  factory _$$EmbeddedDataImplCopyWith(
-          _$EmbeddedDataImpl value, $Res Function(_$EmbeddedDataImpl) then) =
-      __$$EmbeddedDataImplCopyWithImpl<$Res>;
-  @override
-  @useResult
-  $Res call(
-      {int? id,
-      String resource,
-      Map<String, dynamic> attributes,
-      ToolbarConfig toolbar,
-      Metadata metadata});
-
-  @override
-  $ToolbarConfigCopyWith<$Res> get toolbar;
-  @override
-  $MetadataCopyWith<$Res> get metadata;
-}
-
-/// @nodoc
-class __$$EmbeddedDataImplCopyWithImpl<$Res>
-    extends _$EmbeddedDataCopyWithImpl<$Res, _$EmbeddedDataImpl>
-    implements _$$EmbeddedDataImplCopyWith<$Res> {
-  __$$EmbeddedDataImplCopyWithImpl(
-      _$EmbeddedDataImpl _value, $Res Function(_$EmbeddedDataImpl) _then)
-      : super(_value, _then);
-
-  /// Create a copy of EmbeddedData
-  /// with the given fields replaced by the non-null parameter values.
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? id = freezed,
-    Object? resource = null,
-    Object? attributes = null,
-    Object? toolbar = null,
-    Object? metadata = null,
-  }) {
-    return _then(_$EmbeddedDataImpl(
-      id: freezed == id
-          ? _value.id
-          : id // ignore: cast_nullable_to_non_nullable
-              as int?,
-      resource: null == resource
-          ? _value.resource
-          : resource // ignore: cast_nullable_to_non_nullable
-              as String,
-      attributes: null == attributes
-          ? _value._attributes
-          : attributes // ignore: cast_nullable_to_non_nullable
-              as Map<String, dynamic>,
-      toolbar: null == toolbar
-          ? _value.toolbar
-          : toolbar // ignore: cast_nullable_to_non_nullable
-              as ToolbarConfig,
-      metadata: null == metadata
-          ? _value.metadata
-          : metadata // ignore: cast_nullable_to_non_nullable
-              as Metadata,
-    ));
-  }
-}
-
-/// @nodoc
-
-@JsonSerializable(explicitToJson: true)
-class _$EmbeddedDataImpl extends _EmbeddedData {
-  const _$EmbeddedDataImpl(
-      {this.id,
-      required this.resource,
-      final Map<String, dynamic> attributes = const {},
-      this.toolbar = const ToolbarConfig(),
-      this.metadata = const Metadata()})
-      : _attributes = attributes,
-        super._();
-
-  factory _$EmbeddedDataImpl.fromJson(Map<String, dynamic> json) =>
-      _$$EmbeddedDataImplFromJson(json);
-
-  @override
-  final int? id;
-  @override
-  final String resource;
-  final Map<String, dynamic> _attributes;
-  @override
-  @JsonKey()
-  Map<String, dynamic> get attributes {
-    if (_attributes is EqualUnmodifiableMapView) return _attributes;
-    // ignore: implicit_dynamic_type
-    return EqualUnmodifiableMapView(_attributes);
-  }
-
-  @override
-  @JsonKey()
-  final ToolbarConfig toolbar;
-  @override
-  @JsonKey()
-  final Metadata metadata;
-
-  @override
-  String toString() {
-    return 'EmbeddedData(id: $id, resource: $resource, attributes: $attributes, toolbar: $toolbar, metadata: $metadata)';
-  }
-
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _$EmbeddedDataImpl &&
-            (identical(other.id, id) || other.id == id) &&
-            (identical(other.resource, resource) ||
-                other.resource == resource) &&
-            const DeepCollectionEquality()
-                .equals(other._attributes, _attributes) &&
-            (identical(other.toolbar, toolbar) || other.toolbar == toolbar) &&
-            (identical(other.metadata, metadata) ||
-                other.metadata == metadata));
-  }
-
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  @override
-  int get hashCode => Object.hash(runtimeType, id, resource,
-      const DeepCollectionEquality().hash(_attributes), toolbar, metadata);
-
-  /// Create a copy of EmbeddedData
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  @override
-  @pragma('vm:prefer-inline')
-  _$$EmbeddedDataImplCopyWith<_$EmbeddedDataImpl> get copyWith =>
-      __$$EmbeddedDataImplCopyWithImpl<_$EmbeddedDataImpl>(this, _$identity);
-
-  @override
-  Map<String, dynamic> toJson() {
-    return _$$EmbeddedDataImplToJson(
-      this,
-    );
-  }
-}
-
-abstract class _EmbeddedData extends EmbeddedData {
-  const factory _EmbeddedData(
-      {final int? id,
-      required final String resource,
-      final Map<String, dynamic> attributes,
-      final ToolbarConfig toolbar,
-      final Metadata metadata}) = _$EmbeddedDataImpl;
-  const _EmbeddedData._() : super._();
-
-  factory _EmbeddedData.fromJson(Map<String, dynamic> json) =
-      _$EmbeddedDataImpl.fromJson;
-
-  @override
-  int? get id;
-  @override
-  String get resource;
-  @override
-  Map<String, dynamic> get attributes;
-  @override
-  ToolbarConfig get toolbar;
-  @override
-  Metadata get metadata;
-
-  /// Create a copy of EmbeddedData
-  /// with the given fields replaced by the non-null parameter values.
-  @override
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  _$$EmbeddedDataImplCopyWith<_$EmbeddedDataImpl> get copyWith =>
-      throw _privateConstructorUsedError;
-}
-
-ToolbarConfig _$ToolbarConfigFromJson(Map<String, dynamic> json) {
-  return _ToolbarConfig.fromJson(json);
-}
-
-/// @nodoc
-mixin _$ToolbarConfig {
-  String? get titleL10n => throw _privateConstructorUsedError;
-  bool get showToolbar => throw _privateConstructorUsedError;
-
-  /// Serializes this ToolbarConfig to a JSON map.
-  Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
-
-  /// Create a copy of ToolbarConfig
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  $ToolbarConfigCopyWith<ToolbarConfig> get copyWith =>
-      throw _privateConstructorUsedError;
-}
-
-/// @nodoc
-abstract class $ToolbarConfigCopyWith<$Res> {
-  factory $ToolbarConfigCopyWith(
-          ToolbarConfig value, $Res Function(ToolbarConfig) then) =
-      _$ToolbarConfigCopyWithImpl<$Res, ToolbarConfig>;
-  @useResult
-  $Res call({String? titleL10n, bool showToolbar});
-}
-
-/// @nodoc
-class _$ToolbarConfigCopyWithImpl<$Res, $Val extends ToolbarConfig>
-    implements $ToolbarConfigCopyWith<$Res> {
-  _$ToolbarConfigCopyWithImpl(this._value, this._then);
-
-  // ignore: unused_field
-  final $Val _value;
-  // ignore: unused_field
-  final $Res Function($Val) _then;
-
-  /// Create a copy of ToolbarConfig
-  /// with the given fields replaced by the non-null parameter values.
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? titleL10n = freezed,
-    Object? showToolbar = null,
-  }) {
-    return _then(_value.copyWith(
-      titleL10n: freezed == titleL10n
-          ? _value.titleL10n
-          : titleL10n // ignore: cast_nullable_to_non_nullable
-              as String?,
-      showToolbar: null == showToolbar
-          ? _value.showToolbar
-          : showToolbar // ignore: cast_nullable_to_non_nullable
-              as bool,
-    ) as $Val);
-  }
-}
-
-/// @nodoc
-abstract class _$$ToolbarConfigImplCopyWith<$Res>
-    implements $ToolbarConfigCopyWith<$Res> {
-  factory _$$ToolbarConfigImplCopyWith(
-          _$ToolbarConfigImpl value, $Res Function(_$ToolbarConfigImpl) then) =
-      __$$ToolbarConfigImplCopyWithImpl<$Res>;
-  @override
-  @useResult
-  $Res call({String? titleL10n, bool showToolbar});
-}
-
-/// @nodoc
-class __$$ToolbarConfigImplCopyWithImpl<$Res>
-    extends _$ToolbarConfigCopyWithImpl<$Res, _$ToolbarConfigImpl>
-    implements _$$ToolbarConfigImplCopyWith<$Res> {
-  __$$ToolbarConfigImplCopyWithImpl(
-      _$ToolbarConfigImpl _value, $Res Function(_$ToolbarConfigImpl) _then)
-      : super(_value, _then);
-
-  /// Create a copy of ToolbarConfig
-  /// with the given fields replaced by the non-null parameter values.
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? titleL10n = freezed,
-    Object? showToolbar = null,
-  }) {
-    return _then(_$ToolbarConfigImpl(
-      titleL10n: freezed == titleL10n
-          ? _value.titleL10n
-          : titleL10n // ignore: cast_nullable_to_non_nullable
-              as String?,
-      showToolbar: null == showToolbar
-          ? _value.showToolbar
-          : showToolbar // ignore: cast_nullable_to_non_nullable
-              as bool,
-    ));
-  }
-}
-
-/// @nodoc
-
-@JsonSerializable(explicitToJson: true)
-class _$ToolbarConfigImpl implements _ToolbarConfig {
-  const _$ToolbarConfigImpl({this.titleL10n, this.showToolbar = false});
-
-  factory _$ToolbarConfigImpl.fromJson(Map<String, dynamic> json) =>
-      _$$ToolbarConfigImplFromJson(json);
-
-  @override
-  final String? titleL10n;
-  @override
-  @JsonKey()
-  final bool showToolbar;
-
-  @override
-  String toString() {
-    return 'ToolbarConfig(titleL10n: $titleL10n, showToolbar: $showToolbar)';
-  }
-
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _$ToolbarConfigImpl &&
-            (identical(other.titleL10n, titleL10n) ||
-                other.titleL10n == titleL10n) &&
-            (identical(other.showToolbar, showToolbar) ||
-                other.showToolbar == showToolbar));
-  }
-
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  @override
-  int get hashCode => Object.hash(runtimeType, titleL10n, showToolbar);
-
-  /// Create a copy of ToolbarConfig
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  @override
-  @pragma('vm:prefer-inline')
-  _$$ToolbarConfigImplCopyWith<_$ToolbarConfigImpl> get copyWith =>
-      __$$ToolbarConfigImplCopyWithImpl<_$ToolbarConfigImpl>(this, _$identity);
-
-  @override
-  Map<String, dynamic> toJson() {
-    return _$$ToolbarConfigImplToJson(
-      this,
-    );
-  }
-}
-
-abstract class _ToolbarConfig implements ToolbarConfig {
-  const factory _ToolbarConfig(
-      {final String? titleL10n, final bool showToolbar}) = _$ToolbarConfigImpl;
-
-  factory _ToolbarConfig.fromJson(Map<String, dynamic> json) =
-      _$ToolbarConfigImpl.fromJson;
-
-  @override
-  String? get titleL10n;
-  @override
-  bool get showToolbar;
-
-  /// Create a copy of ToolbarConfig
-  /// with the given fields replaced by the non-null parameter values.
-  @override
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  _$$ToolbarConfigImplCopyWith<_$ToolbarConfigImpl> get copyWith =>
-      throw _privateConstructorUsedError;
-}
